% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv2df.R
\name{cv2df}
\alias{cv2df}
\title{Read Zahner CV data}
\usage{
cv2df(datafilename, wearea = NA)
}
\arguments{
\item{datafilename}{complete path to csv file}

\item{wearea}{defaults to NA}
}
\value{
dataframe with data and metadata
}
\description{
This function converts Zahner CV (exported as csv)
to dataframe while retaining metadata.
Supports getting WE area from either datafile (metadata field) or
function argument. The datafile value takes priority.
}

