% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2df.R
\name{txt2df}
\alias{txt2df}
\title{Import Zahner impedance series data}
\usage{
txt2df(datafilename, wearea = NA, runid.width = 3)
}
\arguments{
\item{datafilename}{path to data file (string)}

\item{wearea}{WE area (numeric)}

\item{runid.width}{width of the runid (numeric)}
}
\value{
dataframe with experimental data and metadata
   LOGF               : log(frequency) / Hz
   LOGO               : log(Omega)
   LOGZm              : log(Z) measured / Ohm
   Phim               : Phase / degrees
   Real               : Real part (of impedance)
   Imag               : Imaginary part (of impedance)
   Phis               : Smoothed phase / degrees
   LOGZs              : Smoothed log(Z)
   LOGZzhit           : log(Z) ZHIT / Ohm
   sampleid           : sampleid
   runid              : runid
   Current            : current (metadata)
   Potential          : potential during measurement
   exp.started        : start time
   exp.ended          : end time
   exp.length.seconds : duration in seconds
}
\description{
This function imports Zahner (Thales) impedance data
that was converted using the Bin2Txt.exe utility.
}

