% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eis2df.R
\name{eis2df}
\alias{eis2df}
\title{Read Zahner EIS data}
\usage{
eis2df(datafilename, wearea = NA)
}
\arguments{
\item{datafilename}{complete path to csv file}

\item{wearea}{defaults to NA}
}
\value{
dataframe
}
\description{
This function converts Zahner impedance data
to dataframe while retaining metadata.
}

